Move the tests out of the seesaw package
They were shipped inside the installed package, so every seesaw install
carried its own test suite. The tests/ directory already existed for the
end-to-end scripts, so put them there and make it a package.

Only the BaseTestCase import path changes; the test bodies are untouched.
