+++ tests/runner_test.py
@@ -1,7 +1,7 @@
 from seesaw.pipeline import Pipeline
 from seesaw.runner import SimpleRunner
 from seesaw.task import PrintItem, SimpleTask
-from seesaw.test_base import BaseTestCase
+from tests.test_base import BaseTestCase
 
 class RunnerTest(BaseTestCase):
     def setUp(self):
